==> 10_code/00_final_cleaning.ipynb <==
# Twitter dataset

import pandas as pd
import nltk
from nltk.corpus import stopwords
import re
import string


df= pd.read_csv("/Users/sarwaridas/Downloads/trainingandtestdata/training.1600000.processed.noemoticon.csv",encoding='latin-1',header=None)#,names=["label","ID","date","query","user","text"])
#df.drop(['query','user','ID','date'],axis=1,inplace=True)   
df.head()

df[0]=df[0].replace(to_replace=4,value=1)
df[0].value_counts()

#df.sample(50000).to_csv("sentiment140-small.csv", header=None, index=None)


df[0].unique()

df.columns= ["label","ID","date","query","user","text"]

#setting label values to 0 & 1
# df['label']= df['label'].replace({4:1})
# df.label.value_counts()

# data_pos= df[df['label']==1] #splitting by pos & neg tweets
# data_neg= df[df['label']==0]
# data_pos= data_pos.iloc[:int(20000)] #taking 1/4th of data
# data_neg= data_neg.iloc[:int(20000)]
# data=pd.concat([data_pos,data_neg])

df.label.value_counts()

punc=string.punctuation
def cleaning_punctuations(text):
  translator=str.maketrans("","",punc)
  return text.translate(translator)


def pre_processing_01(tweet):
  '''
  Basic pre-processing to clean data:
  1. Usernames
  2. URLs
  3. Special Characters
  4. Multiple Spaces
  5. Emails
  6. Numbers
  7. Single Chars
  '''
  tweet=tweet.str.lower()
  tweet = tweet.apply(lambda x:re.sub('@[^\s]+','',str(x))) # Remove Handles (aka usernames)
  tweet = tweet.apply(lambda x:re.sub('((www\.[^\s]+)|(https?://[^\s]+))','',str(x))) # Remove URLs
  tweet= tweet.apply(lambda x:' '.join(re.findall(r'\w+', str(x)))) #remove special chars
  #tweet= tweet.apply(lambda x:cleaning_punctuations(x)) #remove punctuations
  tweet = tweet.apply(lambda x:re.sub('@[^\s]+','',str(x))) # Remove emails
  tweet = tweet.apply(lambda x:re.sub('[0-9]+','',str(x))) # Remove numbers
  #
  # tweet= tweet.apply(lambda x:re.sub(r'\s+', ' ', str(x), flags=re.I)) #replace multiple spaces with single space
  tweet = tweet.apply(lambda x:re.sub(r'\s+[a-zA-Z]\s+', '', x))
  return tweet

df['text']= pre_processing_01(df.text)

df.text.sample(20)

from nltk.tokenize import RegexpTokenizer

stopwords_list=set(stopwords.words('english'))
def cleaning_stopwords(text):
    return " ".join([word for word in str(text).split() if word not in stopwords_list])


def pre_processing_02(data):
    '''
    1. Removing stopwords (using NLTK)
    2. Stemming
    3. Lemmitization
    
    '''
    data.text=data.text.apply(lambda x: cleaning_stopwords(x)) #removing stopwords
    # tokenizer=RegexpTokenizer(r'\w+')
    # data.text=data.text.apply(tokenizer.tokenize)
    #data.text= data.text.apply(lambda x: stemming_on_text(x))
    #data.text= data.text.apply(lambda x: lemmatizer_on_text(x))
    return data



#data['text2']= pre_processing_02(data)


#Tokenizing

#Removing stop words

stopwords_list=set(stopwords.words('english'))
def cleaning_stopwords(text):
  return " ".join([word for word in str(text).split() if word not in stopwords_list])
df.text=df.text.apply(lambda x: cleaning_stopwords(x))

# Stemming

def stemming_on_text(data):
    st=nltk.PorterStemmer()
    text= [st.stem(word) for word in data]
    return data
data.text= data.text.apply(lambda x: stemming_on_text(x))

#Lemmitization

from nltk.stem import WordNetLemmatizer 
def lemmatizer_on_text(data):
    wnl = WordNetLemmatizer()
    text= [wnl.lemmatize(word) for word in data]
    return data
data.text= data.text.apply(lambda x: lemmatizer_on_text(x))

all_text = ''.join([c for c in data.text])
words_split = all_text.split(" ")
len(list(set(words_split)))
#print ('Number of words :', len(words_split))

# from collections import Counter
# # Count all the words using Counter Method
# count_words = Counter(words_split)

# total_words = len(count_words)
# sorted_words = count_words.most_common(50000)
# keepwords= list(set([x[0] for x in sorted_words]))
nltk.download('words')
from nltk.corpus import words

def keeping_common_words(text):
    return " ".join([word for word in str(text).split() if word in words.words()])

# data.text=data.text.apply(lambda x: keeping_common_words(x))
# data.text.head()

data.shape

data.to_csv("CLEANED_FOR_REAL.csv")

data.isna().sum()

os.getcwd()
